refactor(frontend): update pages folders
