+++ frontend/src/pages/Authentication/AuthLayout/AuthLayout.jsx
